fix(BaseModel): 将ID字段类型修改为uint64以支持更大的值

+++ internal/models/basemodel.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BaseModel struct {
-	ID        uint           `json:"id" gorm:"primarykey"`
+	ID        uint64         `json:"id" gorm:"primarykey"`
 	CreatedAt time.Time      `json:"-"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"-" gorm:"index"`
